# lms-server/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LMS API
  description: |
    Multi-tenant Learning Management System API
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Multi-Tenancy
    All resources are scoped by tenant. The tenant ID is automatically extracted from:
    - `x-tenant-id` header, or
    - The authenticated user's tenant
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    - Tenant signup: 2 requests per hour
  version: 1.0.0
  contact:
    name: API Support
    email: edinmavric10@gmail.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server (example)
...
paths:
...
  /enrollments:
    get:
      tags:
        - Enrollments
      summary: Get all enrollments
      operationId: getAllEnrollments
      security:
        - bearerAuth: []
      parameters:
        - name: student
          in: query
          schema:
            type: string
        - name: course
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled, paused]
      responses:
        '200':
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enrollment'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Enrollments
      summary: Create enrollment
      description: Enroll student in course (admin or professor only)
      operationId: createEnrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student
                - course
              properties:
                student:
                  type: string
                  description: User ID with student role
                course:
                  type: string
                  description: Course ID
                status:
                  type: string
                  enum: [active, completed, cancelled, paused]
                  default: active
      responses:
        '201':
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /enrollments/{id}:
    get:
      tags:
        - Enrollments
      summary: Get enrollment by ID
      operationId: getEnrollmentById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      responses:
        '200':
          description: Enrollment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Enrollments
      summary: Update enrollment
      operationId: updateEnrollment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, completed, cancelled, paused]
      responses:
        '200':
          description: Enrollment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Enrollments
      summary: Delete enrollment (soft delete)
      description: Soft delete enrollment (admin only)
      operationId: deleteEnrollment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      responses:
        '200':
          description: Enrollment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /enrollments/{id}/payments:
    post:
      tags:
        - Enrollments
      summary: Add payment to enrollment
      description: Record payment for enrollment (admin or professor only)
      operationId: addPayment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EnrollmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  minimum: 0
                  description: Payment amount (cannot exceed course price)
                date:
                  type: string
                  format: date-time
                  default: Current date
      responses:
        '200':
          description: Payment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    EnrollmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Enrollment ID
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Enrollment:
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        student:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
        course:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
                price:
                  type: number
        status:
          type: string
          enum: [active, completed, cancelled, paused]
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        isDeleted:
          type: boolean
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    Payment:
      type: object
      properties:
        amount:
          type: number
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [paid, pending]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'